brace脚本基本运算部分

+++ CppDsl/DslData.cpp
@@ -489,7 +489,7 @@ namespace DslData
         m_FileName = file;
         FILE* fp = fopen(file, "rb");
         if (0 != fp) {
-            bufferSize = fread(buffer, 1, 0x80000, fp);
+            bufferSize = static_cast<int>(fread(buffer, 1, 0x80000, fp));
             fclose(fp);
         }
         LoadBinaryCode(buffer, bufferSize);
